+++ library/src/main/java/com/getstream/sdk/chat/rest/controller/TokenAuthInterceptor.java
@@ -2,6 +2,7 @@ package com.getstream.sdk.chat.rest.controller;
 
 import android.util.Log;
 
+import com.getstream.sdk.chat.StreamChat;
 import com.getstream.sdk.chat.interfaces.CachedTokenProvider;
 import com.getstream.sdk.chat.rest.response.ErrorResponse;
 
@@ -40,7 +41,7 @@ public class TokenAuthInterceptor implements Interceptor {
         try {
             latch.await();
         } catch (InterruptedException e) {
-            e.printStackTrace();
+            StreamChat.getLogger().logT(this, e);
         }
 
         Request request = chain.request();
@@ -51,7 +52,7 @@ public class TokenAuthInterceptor implements Interceptor {
         if (!response.isSuccessful()) {
             ErrorResponse err = ErrorResponse.parseError(response);
             if (err.getCode() == ErrorResponse.TOKEN_EXPIRED_CODE) {
-                Log.d(TAG, "Retrying new request");
+                StreamChat.getLogger().logD(this,"Retrying new request");
                 token = null; // invalidate local cache
                 tokenProvider.tokenExpired();
                 response.close();