+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Support auto capitalization for keyboard
 - Update `UpdateChannelRequest` for reserved fields
 - renamed `MoreActionDialog` to `MessageMoreActionDialog`
+- Add `StreamLoggerHandler` interface for custom logging client data
+- Add logging customization ability
 - fix markdown for mention if there is no space at prefix @
 - fix Edit Attachment behavior
 - add support for channel.hide with clear history + events